Kernel script transformed into a Gui script
The basic plumbing has been set up so that the script can be activated
as a Gui script. The mechanics are exactly the same, because the Gui
script imports the kernel script as a module and then call its
functions. This opens the door to create an interface to provide more
parameters and settings to the remeshing algorithm.