Connection between GUI and remeshing algorithm
A rudimentary connection has been set to be able to activate the
remeshing scrpit from within the GUI. It uses the 'sendCommand' function
to construct a command on the go. This is not the preferred way when
developing an Abaqus GUI script. The best approach is to construct the
command with keyword and let the mode issue it. This will refined in
further commits.