Completed initial GUI for model and part selection
The main interface for selecting models and parts have been (almost)
completed. The selected model and part are associated with AFXKeywords
which use the FX machinery too keep up to date with the GUI elements.
Also, queries have been registered so that changes in the model tree
trigger updates on the GUI. A bug has been found, however, which makes
Abaqus to crash when a model is deleted when this Gui plugin is open.
The error has been documented for future reference. Also, a new file was
added to provide a graphical scheme of the interconnection of the
elements that compose the structure of the plugin.