Integration of the GUI and the kernel script
Using the Gui Command interface, the selections done in the GUI were
connected to the parameters required by the kernel script to run the
algorithm. The kernel script if fully runnable from the GUI. Also,
another GUI feature has been added, by which a check box can be toggled
to match the origin model with the destiny model.