Added example tab together with example inps
The 'examples' tab has been adeed to the plugin GUI. A set of inps have
been linked to the application so it is possible to select an example
and run the remesher algorithm on it. Each example has an image
associated. During the loading, the application parses the inps to get
the part names and description of the parts. A new set of keywords has
been generated to deal with the example parts and models submitted to the
command. This allows us to use the same interface to the remesher
algorithm either from the main tab or from the examples tab.